backend_django/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
asgiref==3.8.1
bcrypt==4.0.1
cryptography==42.0.8
Authlib==1.3.2
oauthlib==3.2.2
aiohttp==3.10.5
httpx==0.27.0
pandas==2.2.2
jsonschema==4.23.0
huggingface-hub==0.22.2
openai==1.42.0
langflow==1.0.16
fastapi==0.111.1
psycopg==3.1.9
gunicorn==22.0.0